[examples] regenerated zips updated to latest version of Xtend Compiler

+++ examples/org.eclipse.xtext.xtext.ui.examples/contents/org.eclipse.xtext.example.xbase.tutorial/src-gen/Xbase01_Literals.java
@@ -7,7 +7,7 @@ public class Xbase01_Literals {
 	public static void main(String[] args) {
 		
 		{/*null*/;/*"Hello World"*/;/*42*/;/*true*/;/*false*/;/*java.lang.String.class*/;
-		  CollectionLiterals.<Integer>newArrayList(3, 5, 7);
+		  CollectionLiterals.<Integer>newArrayList(((Integer)3), ((Integer)5), ((Integer)7));
 		  CollectionLiterals.<String>newHashSet("foo", "bar", "baz");
 		  Pair<Integer,String> _operator_mappedTo = ObjectExtensions.<Integer, String>operator_mappedTo(1, "one");
 		  Pair<Integer,String> _operator_mappedTo_1 = ObjectExtensions.<Integer, String>operator_mappedTo(2, "two");
